add install scripts for next3/next4 utils
before running next4 script install_next4.sh should
be run to build and install the .next4 utils.

the .next3 utils are used for backward compatibility with next3.

I added the next4 script to this repository, so the install scripts
will install it as well.
The same script is used for both next4 and next3.